etcdserver: move unti out of server.go
etcdserver file is messy enough. Let's make it be less messy.
